pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "avrae-ls"
version = "0.1.0"
description = "Language server for Avrae draconic aliases"
authors = [
  { name = "1drturtle" }
]
license-files = ["LICENSE"]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygls>=1.3.1",
    "lsprotocol>=2023.0.1",
    "httpx>=0.27",
    "draconic-interpreter @ git+https://github.com/avrae/draconic@master",
    "d20>=1.1.2",
]

[project.optional-dependencies]
dev = ["pytest>=8.3", "pytest-asyncio>=0.23", "ruff>=0.6"]

[project.scripts]
avrae-ls = "avrae_ls.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120
target-version = "py311"
exclude = [
    "avrae",
]
